% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafonction.R
\encoding{UTF-8}
\name{fix_missing}
\alias{fix_missing}
\title{Recode une valeur en valeur manquante}
\usage{
fix_missing(x, na.value)
}
\arguments{
\item{x}{Un vecteur ou une liste.}

\item{na.value}{La valeur que l'on souhaite recoder en valeur manquante (\code{NA}).}
}
\value{
Un vecteur si \code{x} est un vecteur, une liste si \code{x} est une liste. La
longueur de l'objet renvoyé est identique à la longueur de \code{x}.
}
\description{
Cette fonction permet de recoder une valeur déclarée par l'utilisateur en
valeur manquante (\code{NA}).
}
\examples{
data <- c(10, 20, -99)
fix_missing(data, -99)
}
