% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafonction_df.R
\encoding{UTF-8}
\name{fix_missing_df}
\alias{fix_missing_df}
\title{Recode une valeur en valeur manquante dans un data frame}
\usage{
fix_missing_df(x, na.value)
}
\arguments{
\item{x}{Un \code{data.frame}.}

\item{na.value}{La valeur que l'on souhaite recoder en valeur manquante (\code{NA}).}
}
\value{
Un objet de classe \code{data.frame}.
}
\description{
La fonction \code{fix_missing_df()} permet de recoder une valeur
manquante dans l'ensemble d'un \code{data.frame}.
}
\examples{
df <- data.frame(a = c(12, 23, -99),
                 b = c(-99, 45, 56))
fix_missing_df(df, -99)
}
